pyproject.toml
[project]
name = "xNav-data-utils"
version = "0.1.0"
description = "Processing utilities for xNav data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ffmpeg>=1.4",
    "json-numpy>=2.1.1",
    "lerobot==0.3.3",
    "matplotlib>=3.10.6",
    "numpy>=2.2.6",
    "open3d>=0.19.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["utils"]

[dependency-groups]
base = [
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
]
dev = [
    "ipykernel>=6.30.1",
    "jupyterlab>=4.4.9",
    "notebook>=7.4.7",
    "py-spy>=0.4.1",
    "scipy>=1.15.3",
]
